refactor(rules-builder): extract RuleTypePicker state to factory
